rework quests

+++ src/main/java/me/nils/everhunt/listeners/QuestListener.java
@@ -30,7 +30,7 @@ import org.bukkit.inventory.ItemStack;
 import java.util.ArrayList;
 import java.util.List;
 
-public class QuestListener implements Listener {
+public class QuestListener implements Listener { // TODO add task text
     @EventHandler
     public void onNPCInteract(PlayerInteractAtEntityEvent event) {
         Player player = event.getPlayer();
@@ -41,7 +41,7 @@ public class QuestListener implements Listener {
         if (data != null) {
             if (data.getType().equals(MobType.NPC)) {
                 if (data.getDisplayName().equals("Old Man Dave")) {
-                    if (!(QuestData.getDone(uuid, 1))) {
+                    if (!(QuestData.getDone(uuid, 1)) && !QuestData.hasOngoing(uuid)) {
                         if (QuestData.getCompletion(uuid, 1) == 0) {
                             player.getInventory().addItem(WeaponManager.items.get("Wooden Bat").getItemStack());
                             Chat.npc(player,"Old Man Dave","Can you please kill the springers upstairs?");
@@ -50,7 +50,7 @@ public class QuestListener implements Listener {
                             for (int i = 0; i < 3; i++) {
                                 new Springer(loc);
                             }
-                            QuestData.setCompletion(uuid, 1, 1);
+                            QuestData.setCompletion(player, 1, 1,"Kill the 3 Springers");
                             return;
                         }
                         if (QuestData.getCompletion(uuid, 1) < 4 && QuestData.getCompletion(uuid, 1) >= 1) {
@@ -68,26 +68,25 @@ public class QuestListener implements Listener {
                     }
                 }
                 if (data.getDisplayName().equals("Marcus")) {
-                    if (!(QuestData.getDone(uuid, 2))) {
+                    if (!(QuestData.getDone(uuid, 2)) && !QuestData.hasOngoing(uuid)) {
                         if (QuestData.getCompletion(uuid, 2) == 0) {
                             Chat.npc(player,"Marcus","My wife divorced me after she said that I will");
                             Chat.npc(player,"Marcus","NEVER be a successful weapon smith,");
                             Chat.npc(player,"Marcus","so to prove her wrong I am gonna make the");
                             Chat.npc(player,"Marcus","BEST SWORD EVER I just need a handle can you give me one?");
 
-                            QuestData.setCompletion(uuid, 2, 1);
+                            QuestData.setCompletion(player, 2, 1,"Talk to Marcus while holding the bat");
                             return;
                         }
                         if (QuestData.getCompletion(uuid, 2) >= 1 && Condition.isHolding(player, "Wooden Bat", ItemType.WEAPON) && QuestData.getCompletion(uuid, 2) < 5) {
                             Chat.npc(player,"Marcus","May I use that Wooden Bat as a handle?");
-                            Chat.guide(player,"Type yes in chat while holding the bat");
-                            QuestData.setCompletion(uuid, 2, 2);
+                            QuestData.setCompletion(player, 2, 2,"Type yes in chat while holding the bat");
                             return;
                         }
                         if (QuestData.getCompletion(uuid, 2) >= 3 && Condition.isHolding(player, "Kings Bone", ItemType.ITEM) && QuestData.getCompletion(uuid, 2) < 5) {
                             Chat.npc(player,"Marcus","I can work with this.");
                             player.getInventory().remove(player.getInventory().getItemInMainHand());
-                            QuestData.setCompletion(uuid, 2, 5);
+                            QuestData.setCompletion(player, 2, 5,"Talk to Marcus");
                             Chat.npc(player,"Marcus","Talk to me in one second.");
                             return;
                         }
@@ -104,19 +103,17 @@ public class QuestListener implements Listener {
                             return;
                         }
                     }
-                    if (!QuestData.getDone(uuid,5) && QuestData.getDone(uuid,2)) {
+                    if (!QuestData.getDone(uuid,5) && QuestData.getDone(uuid,2) && !QuestData.hasOngoing(uuid)) {
                         if (QuestData.getCompletion(uuid,5) == 0) {
                             Chat.npc(player,"Marcus","I need new materials for Lucia II.");
                             Chat.npc(player,"Marcus","Give me 2 compressed stone pls?");
-                            Chat.guide(player,"find a way to collect compressed stone");
-                            QuestData.setCompletion(uuid,5,1);
+                            QuestData.setCompletion(player,5,1,"Find a way to collect compressed stone");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,5) == 3 && Condition.isHolding(player,"Compressed Stone",ItemType.ITEM,2)) {
                             Chat.npc(player,"Marcus","Thanks for helping me again!");
-                            Chat.guide(player,"click him again while holding Lucia I");
                             player.getInventory().remove(player.getInventory().getItemInMainHand());
-                            QuestData.setCompletion(uuid,5,4);
+                            QuestData.setCompletion(player,5,4,"Click him again while holding Lucia I");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,5) == 4 && Condition.isHolding(player,"Lucia I",ItemType.WEAPON)) {
@@ -163,24 +160,23 @@ public class QuestListener implements Listener {
                         if (QuestData.getCompletion(uuid,3) == 6) {
                             Chat.npc(player,"Mi","Please, go kill a Undead Scarecrow!");
                             Chat.npc(player,"Mi","I would really appreciate it.");
-                            QuestData.setCompletion(uuid,3,7);
+                            QuestData.setCompletion(player,3,7,"Kill 1 Undead Scarecrow");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,3) == 8) {
                             Chat.npc(player,"Mi","Thanks.");
                             player.getInventory().addItem(HelmetManager.items.get("Jester Helmet").getItemStack());
-                            QuestData.setCompletion(uuid,3,9);
+                            QuestData.setCompletion(player,3,9,"Talk to the Guild Master");
                             return;
                         }
                     }
                 }
                 if (data.getDisplayName().equals("Hunter")) {
-                    if (!(QuestData.getDone(uuid, 3)) && QuestData.getDone(uuid,2)) {
+                    if (!(QuestData.getDone(uuid, 3)) && QuestData.getDone(uuid,2) && !QuestData.hasOngoing(uuid)) {
                         if (QuestData.getCompletion(uuid, 3) <= 1) {
                             Chat.npc(player,"Hunter","Are you the hunter who killed the Wolf King?");
                             Chat.npc(player,"Hunter","That's amazing can I have your armor?");
-                            Chat.guide(player,"Type yes while wearing the armor");
-                            QuestData.setCompletion(uuid, 3, 1);
+                            QuestData.setCompletion(player, 3, 1,"Type yes while wearing the armor");
                             return;
                         }
                         if (QuestData.getCompletion(uuid, 3) == 2 && Condition.isHolding(player, "Wheat", ItemType.ITEM, 30)) {
@@ -191,7 +187,7 @@ public class QuestListener implements Listener {
                             player.getInventory().getItemInMainHand().setAmount(amount);
                             player.getInventory().addItem(ArmorManager.items.get("Jester Boots").getItemStack());
                             new TeleportData(uuid, "Guild House", 186, -58, -56);
-                            QuestData.setCompletion(uuid, 3, 3);
+                            QuestData.setCompletion(player, 3, 3,"Go to the Guild House");
                             Chat.guide(player,"Open your warp menu by typing /warp");
                             return;
                         } else if (QuestData.getCompletion(uuid, 3) == 2) {
@@ -203,14 +199,13 @@ public class QuestListener implements Listener {
                             Chat.npc(player,"Hunter","Wow, you look pretty strong!");
                             Chat.npc(player,"Hunter","Here, have this!");
                             player.getInventory().addItem(ArmorManager.items.get("Jester Chestplate").getItemStack());
-                            QuestData.setCompletion(uuid,3,6);
+                            QuestData.setCompletion(player,3,6,"Look for a way to get the helmet");
                             return;
                         } else if (QuestData.getCompletion(uuid,3) == 4) {
                             Chat.npc(player,"Hunter","Want to buy the Jester Leggings for 100 coins?");
                             if (PlayerData.data.get(uuid).pay(100)) {
-                                QuestData.setCompletion(uuid, 3, 5);
+                                QuestData.setCompletion(player, 3, 5,"Talk to him again while holding your strongest weapon");
                                 player.getInventory().addItem(Items.getBase("Jester Leggings"));
-                                Chat.guide(player, "Talk to him again while holding your strongest weapon");
                             } else Chat.guide(player, "To sell items you can use /market to sell to players or you can sell your items with /sell if they have an item value");
                             return;
                         }
@@ -222,8 +217,7 @@ public class QuestListener implements Listener {
                             Chat.npc(player,"Guild Master","Hello, I am the guild master!");
                             Chat.npc(player,"Guild Master","I see that you are gathering the pieces for the Jester set!");
                             Chat.npc(player,"Guild Master","If you talk to me when you have the complete set, I'll officially make you a hunter!");
-                            Chat.guide(player,"Look around the guild house and see if you can get another piece.");
-                            QuestData.setCompletion(uuid,3,4);
+                            QuestData.setCompletion(player,3,4,"Look around the guild house and see if you can get another piece.");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,3) == 9) {
@@ -236,19 +230,19 @@ public class QuestListener implements Listener {
                     }
                 }
                 if (data.getDisplayName().equals("Farmer")) {
-                    if (!QuestData.getDone(uuid,4) && QuestData.getDone(uuid,3)) {
+                    if (!QuestData.getDone(uuid,4) && QuestData.getDone(uuid,3) && !QuestData.hasOngoing(uuid)) {
                         if (Condition.isHolding(player,"Gardening Hoe",ItemType.TOOL) && QuestData.getCompletion(uuid,4) < 1) {
                             Chat.npc(player,"Farmer","Have this backpack now go farm until you have 128 wheat!");
                             Chat.guide(player,"You can open the backpack by holding it and pressing right click on your mouse");
                             player.getInventory().addItem(Items.getBase("Backpack"));
-                            QuestData.setCompletion(uuid,4,1);
+                            QuestData.setCompletion(player,4,1,"Collect 128 wheat");
                             return;
                         } else if (QuestData.getCompletion(uuid,4) < 1) {
                             Chat.guide(player,"Talk to him while holding the gardening hoe");
                         }
                         if (QuestData.getCompletion(uuid,4) == 1 && Condition.has(player,"Wheat",ItemType.ITEM,128)) {
                             Chat.npc(player,"Farmer","Give me your straw hat and i'll upgrade it.");
-                            QuestData.setCompletion(uuid,4,2);
+                            QuestData.setCompletion(player,4,2,"Talk to him again while holding your Straw Hat");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,4) == 2 && Condition.isHolding(player,"Straw Hat",ItemType.ARMOR)) {
@@ -265,16 +259,15 @@ public class QuestListener implements Listener {
                             Chat.npc(player,"Tim","Buy my drill for 500 coins!");
                             Chat.npc(player,"Tim","If you want to save my mining business.");
                             if (PlayerData.data.get(uuid).pay(500)) {
-                                QuestData.setCompletion(uuid, 5, 2);
+                                QuestData.setCompletion(player, 5, 2,"Talk to Tim while holding the drill to proceed");
                                 player.getInventory().addItem(Items.getBase("Drill"));
-                                Chat.guide(player,"talk to tim while holding the drill to proceed");
                             } else Chat.guide(player, "To sell items you can use /market to sell to players or you can sell your items with /sell if they have an item value");
                             return;
                         }
                         if (QuestData.getCompletion(uuid,5) == 2 && Condition.isHolding(player,"Drill",ItemType.TOOL)) {
                             new TeleportData(uuid,"Mine",211,-16,187);
                             Chat.guide(player,"use /warp");
-                            QuestData.setCompletion(uuid,5,3);
+                            QuestData.setCompletion(player,5,3,"Go to the Mine");
                             return;
                         }
                     }
@@ -318,18 +311,18 @@ public class QuestListener implements Listener {
             if (entity.getName().equals("Springer")) {
                 if (!(QuestData.getDone(uuid, 1)) && QuestData.getCompletion(uuid, 1) >= 1) {
                     double completion = QuestData.getCompletion(uuid, 1) + 1;
-                    QuestData.setCompletion(uuid, 1, completion);
+                    QuestData.setCompletion(player, 1, completion,"Kill all the Springers");
                 }
             }
             if (entity.getName().equals("Undead Scarecrow")) {
                 if (!(QuestData.getDone(uuid,3)) && QuestData.getCompletion(uuid,3) == 7) {
-                    QuestData.setCompletion(uuid,3,8);
+                    QuestData.setCompletion(player,3,8,"Talk to Mi");
                 }
             }
             if (entity.getName().equals("Wolf King")) {
                 if (!(QuestData.getDone(uuid, 2)) && QuestData.getCompletion(uuid, 2) >= 1) {
                     double completion = QuestData.getCompletion(uuid, 2) + 2;
-                    if (completion > 4) QuestData.setCompletion(uuid, 2, 4);
+                    if (completion > 4) QuestData.setCompletion(player, 2, 4, "Talk to Mi");
                     PlayerData.data.get(uuid).addXp(50);
                 }
             }
@@ -346,7 +339,7 @@ public class QuestListener implements Listener {
                 PlayerData.data.get(uuid).addXp(25);
                 player.getInventory().remove(player.getInventory().getItemInMainHand());
                 Chat.npc(player,"Marcus","Can you please go fetch me something to make the blade?");
-                QuestData.setCompletion(uuid, 2, 3);
+                QuestData.setCompletion(player, 2, 3,"Look for the Kings Bone");
                 return;
             }
         }
@@ -362,7 +355,7 @@ public class QuestListener implements Listener {
                 Chat.npc(player,"Hunter","Have this!");
                 Chat.npc(player,"Hunter","Come back with 30 wheat");
                 Chat.npc(player,"Hunter","and I'll give you the special boots.");
-                QuestData.setCompletion(uuid, 3, 2);
+                QuestData.setCompletion(player, 3, 2,"Collect 30 wheat");
                 return;
             }
         }
